Rename ufsdat.h to libufsdat.h
This is in prep for creating a new ufsdat.h in the kernel UFS code.
