+++ sys/src/9/port/devufs.c
@@ -20,7 +20,7 @@
 #include "fns.h"
 #include "../port/error.h"
 
-#include "ufs/ufsdat.h"
+#include "ufs/libufsdat.h"
 #include "../ufs/ufs_ext.h"
 
 enum {
