+++ sys/src/9/ufs/ffs_subr.c
@@ -35,7 +35,7 @@
 #include "dat.h"
 #include "fns.h"
 
-#include <ufs/ufsdat.h>
+#include <ufs/libufsdat.h>
 #include <ufs/freebsd_util.h>
 #include "ufs_harvey.h"
