+++ sys/src/9/ufs/ffs_vfsops.c
@@ -37,7 +37,7 @@
 #include "dat.h"
 #include "port/portfns.h"
 
-#include <ufs/ufsdat.h>
+#include <ufs/libufsdat.h>
 #include <ufs/freebsd_util.h>
 #include "ufs_ext.h"
 #include "ufs_mountpoint.h"
