+++ sys/src/9/ufs/ufs_harvey.c
@@ -15,7 +15,7 @@
 #include "dat.h"
 #include "fns.h"
 
-#include <ufs/ufsdat.h>
+#include <ufs/libufsdat.h>
 #include <ufs/freebsd_util.h>
 #include "ufs_harvey.h"
 #include "ufs_mountpoint.h"
