+++ sys/src/9/ufs/ufs_inode.c
@@ -40,7 +40,7 @@
 #include "dat.h"
 #include "port/portfns.h"
 
-#include <ufs/ufsdat.h>
+#include <ufs/libufsdat.h>
 #include <ufs/freebsd_util.h>
 #include "ufs_harvey.h"
