+++ sys/src/9/ufs/ufs_mountpoint.c
@@ -15,7 +15,7 @@
 #include "dat.h"
 #include "fns.h"
 
-#include <ufs/ufsdat.h>
+#include <ufs/libufsdat.h>
 #include <ufs/freebsd_util.h>
 #include <ufs/fs.h>
 #include "ufs_ext.h"
