+++ sys/src/cmd/newfs/newfs.c
@@ -40,7 +40,7 @@
 #include <libc.h>
 
 #include <ctype.h>
-#include <ufs/ufsdat.h>
+#include <ufs/libufsdat.h>
 #include <ufs/fs.h>
 #include <ufs/libufs.h>
 #include "newfs.h"
