+++ sys/src/libufs/ffs_tables.c
@@ -32,7 +32,7 @@
 #include <u.h>
 #include <libc.h>
 
-#include <ufs/ufsdat.h>
+#include <ufs/libufsdat.h>
 #include <ufs/dinode.h>
 #include <ufs/fs.h>
